+++ aa_intel_tool/admin.py
@@ -86,6 +86,7 @@ class ScanAdmin(BaseReadOnlyAdminMixin, admin.ModelAdmin):
         :return:
         :rtype:
         """
+
         intel_type = obj.get_scan_type_display()
         scan_link = reverse_absolute(
             viewname="aa_intel_tool:intel_tool_scan", args=[obj.hash]