+++ aa_intel_tool/constants.py
@@ -11,7 +11,7 @@ from django.utils.text import slugify
 from django.utils.translation import gettext_lazy as _
 
 # AA Intel Tool
-# import aa_intel_tool.parser.module.chatlist
+import aa_intel_tool.parser.module.chatlist
 import aa_intel_tool.parser.module.dscan
 import aa_intel_tool.parser.module.fleetcomp
 from aa_intel_tool import __version__
@@ -48,7 +48,7 @@ class AdditionalEveCategoryId(IntEnum):
 
 # Pre-compiled regex patterns used throughout the app
 REGEX_PATTERN = {
-    # "chatlist": re.compile(pattern=r"(?im)^[a-zA-Z0-9\u0080-\uFFFF -_]{3,37}$"),
+    "chatlist": re.compile(pattern=r"(?im)^[a-zA-Z0-9\u0080-\uFFFF -_]{3,37}$"),
     "dscan": re.compile(
         pattern=rf"(?im)^(\d+)[\t](.*)[\t](.*)[\t](-|(.*) ({distance_units}))$",
         flags=re.VERBOSE,
@@ -67,11 +67,11 @@ REGEX_PATTERN = {
 
 # Supported intel types and their parameters
 SUPPORTED_INTEL_TYPES = {
-    # "chatlist": {
-    #     "name": _("Chat List"),
-    #     "parser": aa_intel_tool.parser.module.chatlist.parse,
-    #     "pattern": REGEX_PATTERN["chatlist"],
-    # },
+    "chatlist": {
+        "name": _("Chat List"),
+        "parser": aa_intel_tool.parser.module.chatlist.parse,
+        "pattern": REGEX_PATTERN["chatlist"],
+    },
     "dscan": {
         "name": _("D-Scan"),
         "parser": aa_intel_tool.parser.module.dscan.parse,