+++ CHANGELOG.md
@@ -9,6 +9,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [In Development] - Unreleased
 
 
+## [0.0.1-alpha.12] - 2023-08-24
+
+### Added
+
+- Upwell structures on grid to D-Scan result page
+- Deployables on grid to D-Scan result page
+- Pos / Pos Modules on grid to D-Scan result page
+- Destination system to Ansiblex Jump Gate in D-Scan
+
+### Changed
+
+- Chat list parser re-enabled. See Warning in [README.md](README.md).
+
+
 ## [0.0.1-alpha.11] - 2023-08-23
 
 ### Changed